#  Optimize 64-bit division-by-constant for x86 platforms