@StephanTLavavej 2018-08-16
This PR continues to have great performance on my i7-4790. With my benchmark changes in #74 (defaulting to "inverted mode" for realistic branch prediction), I observe a 45% improvement for MSVC, 48% for Clang, and larger improvements for smaller output (which performs more divisions to trim more digits):

| MSVC Digits |  Before |   After | Speedup |
| ----------: | :-----: | :-----: | :-----: |
|         All | 153.713 | 106.1   | 1.45    |
|           7 | 439.555 | 167.245 | 2.63    |
|           6 | 476.098 | 178.361 | 2.67    |
|           5 | 510.015 | 186.656 | 2.73    |
|           4 | 545.182 | 195.279 | 2.79    |
|           3 | 584.221 | 206.642 | 2.83    |
|           2 | 635.785 | 221.941 | 2.86    |
|           1 | 711.366 | 238.35  | 2.98    |

| Clang Digits |  Before |   After | Speedup |
| -----------: | :-----: | :-----: | :-----: |
|          All | 118.688 | 80.378  | 1.48    |
|            7 | 284.491 | 130.421 | 2.18    |
|            6 | 305.25  | 137.703 | 2.22    |
|            5 | 336.106 | 145.762 | 2.31    |
|            4 | 356.973 | 153.755 | 2.32    |
|            3 | 379.057 | 158.039 | 2.40    |
|            2 | 412.38  | 177.199 | 2.33    |
|            1 | 486.917 | 185.11  | 2.63    |